+++ CHANGELOG.md
@@ -2,6 +2,106 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [11.8.1](https://github.com/mojaloop/central-services-logger/compare/v11.8.0...v11.8.1) (2025-03-31)
+
+
+### Bug Fixes
+
+* improve stringify and error handling in UdpStream ([#63](https://github.com/mojaloop/central-services-logger/issues/63)) ([1966d1c](https://github.com/mojaloop/central-services-logger/commit/1966d1cc7cebbed34bbdb5df8179a3f9c9ebd83f))
+
+## [11.8.0](https://github.com/mojaloop/central-services-logger/compare/v11.7.0...v11.8.0) (2025-03-26)
+
+
+### Features
+
+* enable passing objects in CSL_LOG_TRANSPORT ([#62](https://github.com/mojaloop/central-services-logger/issues/62)) ([1692da6](https://github.com/mojaloop/central-services-logger/commit/1692da6d3e35a5d150783bfe55d76ac789e59d9c))
+
+## [11.7.0](https://github.com/mojaloop/central-services-logger/compare/v11.6.2...v11.7.0) (2025-03-18)
+
+
+### Features
+
+* enhance logging to handle expected errors and update tests ([#61](https://github.com/mojaloop/central-services-logger/issues/61)) ([9265a4a](https://github.com/mojaloop/central-services-logger/commit/9265a4a76641f8d24754bbd40a78ea13ceb11d7e))
+
+### [11.6.2](https://github.com/mojaloop/central-services-logger/compare/v11.6.1...v11.6.2) (2025-03-12)
+
+
+### Bug Fixes
+
+* update stream import to use node:stream ([#60](https://github.com/mojaloop/central-services-logger/issues/60)) ([dab621a](https://github.com/mojaloop/central-services-logger/commit/dab621ada6ec277acac37e0010489f179d4c8b42))
+
+### [11.6.1](https://github.com/mojaloop/central-services-logger/compare/v11.6.0...v11.6.1) (2025-03-11)
+
+
+### Bug Fixes
+
+* use production friendly default ([#59](https://github.com/mojaloop/central-services-logger/issues/59)) ([74ffcd8](https://github.com/mojaloop/central-services-logger/commit/74ffcd83c97661c40bc67f22a53e11f4cc2e2430))
+
+## [11.6.0](https://github.com/mojaloop/central-services-logger/compare/v11.5.5...v11.6.0) (2025-03-11)
+
+
+### Features
+
+* multiple configurable transports ([#58](https://github.com/mojaloop/central-services-logger/issues/58)) ([cd24562](https://github.com/mojaloop/central-services-logger/commit/cd24562d86a1ced7bb23b31817327e1badd69189))
+
+### [11.5.5](https://github.com/mojaloop/central-services-logger/compare/v11.5.4...v11.5.5) (2025-02-18)
+
+### [11.5.4](https://github.com/mojaloop/central-services-logger/compare/v11.5.3...v11.5.4) (2025-01-27)
+
+
+### Bug Fixes
+
+* add missing types ([#56](https://github.com/mojaloop/central-services-logger/issues/56)) ([b1b1c01](https://github.com/mojaloop/central-services-logger/commit/b1b1c011e555f03f5ee9e059593766c372d3393a))
+
+### [11.5.3](https://github.com/mojaloop/central-services-logger/compare/v11.5.2...v11.5.3) (2025-01-24)
+
+### [11.5.2](https://github.com/mojaloop/central-services-logger/compare/v11.5.0...v11.5.2) (2025-01-08)
+
+
+### Bug Fixes
+
+* **csi-528:** fixed logger.child() to avoid creating new winston instance ([#53](https://github.com/mojaloop/central-services-logger/issues/53)) ([3768027](https://github.com/mojaloop/central-services-logger/commit/3768027ac9d9020d664a238eed2fd1595294a5f8))
+
+### [11.5.1](https://github.com/mojaloop/central-services-logger/compare/v11.5.0...v11.5.1) (2024-08-22)
+
+
+### Bug Fixes
+
+* **csi-528:** added husky ([201bce1](https://github.com/mojaloop/central-services-logger/commit/201bce17861fdf0a05d0c80b9e5d4a416b680a22))
+* **csi-528:** fixed logger.child() to avoid creating new winston instance ([963ac80](https://github.com/mojaloop/central-services-logger/commit/963ac802e56d10c7f15777c8dba3060d1f0b5bfc))
+* **csi-528:** updated deps ([d010671](https://github.com/mojaloop/central-services-logger/commit/d0106712d05162c04780882e9422c52bf3b4e70b))
+* **csi-528:** updated deps ([7d39c51](https://github.com/mojaloop/central-services-logger/commit/7d39c51ca63243ae991c0be5fa10eb7eb0392a7e))
+
+## [11.5.0](https://github.com/mojaloop/central-services-logger/compare/v11.3.1...v11.5.0) (2024-07-25)
+
+
+### Features
+
+* **csi-334:** added contextLogger with asyncStorage ([#52](https://github.com/mojaloop/central-services-logger/issues/52)) ([8a46e3e](https://github.com/mojaloop/central-services-logger/commit/8a46e3e8f36f7d18f34f90a5c0883d56c2192123))
+
+### [11.4.5](https://github.com/mojaloop/central-services-logger/compare/v11.4.4...v11.4.5) (2024-07-24)
+
+### [11.4.4](https://github.com/mojaloop/central-services-logger/compare/v11.4.3...v11.4.4) (2024-07-24)
+
+### [11.4.3](https://github.com/mojaloop/central-services-logger/compare/v11.4.2...v11.4.3) (2024-07-24)
+
+### [11.4.2](https://github.com/mojaloop/central-services-logger/compare/v11.4.1...v11.4.2) (2024-07-24)
+
+### [11.4.1](https://github.com/mojaloop/central-services-logger/compare/v11.4.0...v11.4.1) (2024-07-24)
+
+## [11.4.0](https://github.com/mojaloop/central-services-logger/compare/v11.4.0-snapshot.1...v11.4.0) (2024-07-22)
+
+### [11.3.1](https://github.com/mojaloop/central-services-logger/compare/v11.3.0...v11.3.1) (2024-04-24)
+
+## [11.3.0](https://github.com/mojaloop/central-services-logger/compare/v11.2.2...v11.3.0) (2024-03-19)
+
+
+### Features
+
+* avoid stringify if only message passed (no extra quotes) ([#47](https://github.com/mojaloop/central-services-logger/issues/47)) ([c172d5e](https://github.com/mojaloop/central-services-logger/commit/c172d5ece8d10e078dbfe9b251aae6a9c35faadc))
+
+### [11.2.2](https://github.com/mojaloop/central-services-logger/compare/v11.2.1...v11.2.2) (2023-08-29)
+
 ### [11.2.1](https://github.com/mojaloop/central-services-logger/compare/v11.2.0...v11.2.1) (2023-08-24)
 
 ## [11.2.0](https://github.com/mojaloop/central-services-logger/compare/v11.1.0...v11.2.0) (2022-11-21)